pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "keithley_client"
version = "0.1.0"
description = "A client for the Keithley SMU"
authors = [{ name = "Fabio Terranova" }]
license = { file = "LICENSE" }
readme = "README.md"
dependencies = [
	"argparse",
	"numpy",
	"pandas",
	"pyqt5",
	"pyqtgraph",
	"pyvisa-py",
	"pyUSB",
	"gpib-ctypes",
	"matplotlib",
]

[project.scripts]
keithley_client = "keithley_client:cli"

[tool.hatch.envs.default]
dependencies = ["pre-commit"]
